Use size_t
git-svn-id: https://svn.blender.org/svnroot/verse/trunk/verse@44 49c17534-1d02-dc11-b283-0030488c597c

+++ v_network_que.h
@@ -12,7 +12,7 @@ extern void	v_nq_send_ack_nak_buf(VNetQueue *queue, VCMDBufHead *buf);
 extern void	v_nq_send_ack(VNetQueue *queue, unsigned int id);
 extern void	v_nq_send_nak(VNetQueue *queue, unsigned int id);
 
-extern char *	v_nq_get_packed(VNetQueue *queue, unsigned int *length);
-extern void	v_nq_store_packed(VNetQueue *queue, const char *data, unsigned int length);
+extern char *	v_nq_get_packed(VNetQueue *queue, size_t *length);
+extern void	v_nq_store_packed(VNetQueue *queue, const char *data, size_t length);
 
 extern boolean	v_nq_send_queue(VNetQueue *queue, void *address);
